Point CF distribution to bucket regional domain name.
Per https://stackoverflow.com/questions/38735306/aws-cloudfront-redirecting-to-s3-bucket
According to the discussion on AWS Developer Forums: Cloudfront domain redirects to S3 Origin URL, it takes time for DNS records to be created and propagated for newly created S3 buckets. The issue is not visible for buckets created in US East (N. Virginia) region, because this region is the default one (fallback).